+++ lib/ooxml_parser/xlsx_parser/xlsx_data/view_model/workbook/worksheet.rb
@@ -17,6 +17,8 @@ module OoxmlParser
     attr_accessor :relationships
     # @return [Relationships] array of ole objects
     attr_accessor :ole_objects
+    # @return [ExtensionList] list of extensions
+    attr_accessor :extension_list
 
     def initialize
       @columns = []
@@ -102,6 +104,8 @@ module OoxmlParser
           end
         when 'oleObjects'
           worksheet.ole_objects = OleObjects.new(parent: worksheet).parse(worksheet_node_child)
+        when 'extLst'
+          worksheet.extension_list = ExtensionList.new(parent: self).parse(worksheet_node_child)
         end
       end
       worksheet.comments = ExcelComments.parse_file(File.basename(path_to_xml_file), OOXMLDocumentObject.path_to_folder)