+++ lib/ooxml_parser/xlsx_parser/xlsx_data/view_model/workbook/worksheet/table_part/extension_list/extension.rb
@@ -1,9 +1,12 @@
+require_relative 'extension/sparkline_groups'
 require_relative 'extension/x14_table'
 module OoxmlParser
   # Class for `ext` data
   class Extension < OOXMLDocumentObject
     # @return [X14Table] table data in x14 namespace
     attr_accessor :table
+    # @return [SparklineGroups] list of groups
+    attr_reader :sparkline_groups
 
     # Parse Extension data
     # @param [Nokogiri::XML:Element] node with Extension data
@@ -13,6 +16,8 @@ module OoxmlParser
         case column_node.name
         when 'table'
           @table = X14Table.new(parent: self).parse(column_node)
+        when 'sparklineGroups'
+          @sparkline_groups = SparklineGroups.new(parent: self).parse(column_node)
         end
       end
       self